+++ test/test_player_scrabble.py
@@ -67,5 +67,67 @@ class TestPlayer(unittest.TestCase):
         ]
         is_valid = player.has_letters(tiles)
         self.assertEqual(is_valid, False)
+
+
+
+    def test_validate_user_has_letters_dobleletter(self):
+        bag_tile = BagTiles()
+        player = Player(1)
+        player.tiles = [
+            ('O', 1),
+            ('O', 1),
+            ('LL', 1),
+            ('A', 1),
+            ('C', 1),
+            ('U', 1),
+            ('M', 1),]
+        tiles = [
+            ('O', 1),
+            ('L', 1),
+            ('L', 1),
+            ('A', 1),
+        ]
+        is_valid = player.has_letters(tiles)
+        self.assertEqual(is_valid, True)
+
+    def test_validate_user_has_letters_2(self):
+        bag_tile = BagTiles()
+        player = Player(1)
+        player.tiles = [
+            ('O', 1),
+            ('L', 1),
+            ('L', 4),
+            ('A', 1),
+            ('C', 1),
+            ('U', 1),
+            ('M', 1),]
+        tiles = [
+            ('O', 1),
+            ('LL', 1),
+            ('A', 1),
+        ]
+        is_valid = player.has_letters(tiles)
+        self.assertEqual(is_valid, True)
+
+    def test_validate_user_has_not_letters(self):
+        bag_tile = BagTiles()
+        player = Player(1)
+        player.tiles = [
+            ('H', 1),
+            ('O', 1),
+            ('LL', 1),
+            ('A', 1),
+            ('C', 1),
+            ('U', 1),
+            ('M', 1),]
+        tiles = [
+            ('H', 1),
+            ('O', 1),
+            ('L', 1),
+            ('A', 1),
+        ]
+        is_valid = player.has_letters(tiles)
+        self.assertEqual(is_valid, False)
+
 if __name__ == '__main__':
     unittest